pyproject.toml
[project]
name = "boilerplate-django-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular[sidecar]>=0.28.0",
    "gunicorn>=23.0.0",
    "mongoengine>=0.29.1",
    "mysqlclient>=2.2.7",
    "python-json-logger>=3.3.0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "faker>=36.2.2",
    "factory-boy>=3.3.3",
    "coverage>=7.6.12",
    "django-debug-toolbar>=5.0.1 ",
    "pytest-django>=4.10.0",
    "django-stubs[compatible-mypy]>=5.1.3",
    "djangorestframework-stubs[compatible-mypy]>=3.15.3",
    "mongomock>=4.3.0",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
exclude = [
    'django_project',
]
